feat: merge fake list api
