+++ include/GMLIB/Server/FakeListAPI.h
@@ -1,30 +1,31 @@
 #pragma once
-#include "GMLIB/DllExport.h"
+#include "GMLIB/GMLIB.h"
 #include "mc/network/packet/PlayerListEntry.h"
 
-namespace GMLIB::FakeListAPI {
+class GMLIB_FakeList {
+public:
+    GMLIB_API static bool addFakeList(PlayerListEntry entry);
 
-GMLIB_API bool addFakeList(PlayerListEntry entry);
+    GMLIB_API static bool
+    addFakeList(std::string name, std::string xuid, ActorUniqueID uniqueId, mce::UUID uuid = mce::UUID::random());
 
-GMLIB_API bool
-addFakeList(std::string name, std::string xuid, ActorUniqueID uniqueId, mce::UUID uuid = mce::UUID::random());
+    GMLIB_API static bool removeFakeList(std::string nameOrXuid);
 
-GMLIB_API bool removeFakeList(std::string nameOrXuid);
+    GMLIB_API static void removeAllFakeLists();
 
-GMLIB_API void removeAllFakeLists();
+    GMLIB_API static PlayerListEntry getFakeList(std::string name);
 
-GMLIB_API PlayerListEntry getFakeList(std::string name);
+    GMLIB_API static bool checkFakeListExistsName(std::string name);
 
-GMLIB_API bool checkFakeListExistsName(std::string name);
+    GMLIB_API static bool checkFakeListExists(std::string name, std::string xuid);
 
-GMLIB_API bool checkFakeListExists(std::string name, std::string xuid);
+    GMLIB_API static std::vector<std::string> getAllFakeNames();
 
-GMLIB_API std::vector<std::string> getAllFakeNames();
+    GMLIB_API static void setListName(std::string realName, std::string fakeName);
 
-GMLIB_API void replaceList(bool add, std::string oldName, std::string newName);
+    GMLIB_API static void resetListName(std::string realName);
 
-GMLIB_API void setSimulatedPlayerOpt(bool set);
+    GMLIB_API static void setSimulatedPlayerListOptimizeEnabled(bool value = true);
 
-GMLIB_API bool getSimulatedPlayerOpt();
-
-} // namespace GMLIB::FakeListAPI
+    GMLIB_API static bool getSimulatedPlayerListOptimizeEnabled();
+};
